Simplify image related calculations. Improvements.
Before, we were manually calculating values for the page ratios.
There is an easier way, which is to ask the ImageView for its
matrix and use that for all the calculations.

This patch also improves some of the text at the top and bottom of
the pages (for sura name juz', and page number). On some page ratios,
the top may come on top of a the page header on pages which start with
a sura header. The correct fix for this is an updated ayahinfo
database that contains the headers (not just the ayat).

Fixed a crash when downloading a range of suras.