Fix crashes. Fix spinner size bugs.
The spinner was being cut off in the jump dialog on gingerbread. This
should fix that. It also lets long press on the translation view always
copy the verse (instead of allowing a contextual action bar on
honeycomb, which is currently broken due to a bug in appcompat). Also
fix some other minor issues.